+++ backend/freelancePM/src/main/java/com/freelance/freelancepm/service/InvoicePdfService.java
@@ -18,17 +18,17 @@ public class InvoicePdfService {
     private final InvoiceRepository invoiceRepository;
     private final InvoiceLineItemRepository invoiceLineItemRepository;
 
-    public byte[] generateInvoicePdf(Long invoiceId) {
+    public byte[] generateInvoicePdf(Integer invoiceId) {
         // Fetch Invoice by ID
         Invoice invoice = invoiceRepository.findById(invoiceId)
                 .orElseThrow(() -> new IllegalArgumentException("Invoice not found with ID: " + invoiceId));
-        
+
         // Fetch Line items
         List<InvoiceLineItem> lineItems = invoiceLineItemRepository.findByInvoiceId(invoiceId);
-        
+
         // TODO: Implement PDF layout generation using Apache PDFBox
         log.info("Generating PDF for invoice ID: {} with {} line items", invoiceId, lineItems.size());
-        
+
         // Return placeholder byte[] for now
         return new byte[0];
     }
